wiring state to request hooks

+++ java/edu/ualr/bittorrent/impl/core/messages/CancelImpl.java
@@ -16,15 +16,15 @@ public class CancelImpl extends PeerMessage<Cancel> implements Cancel {
     this.length = length;
   }
 
-  public long getBeginningOffset() {
+  public int getBeginningOffset() {
     return begin;
   }
 
-  public long getBlockLength() {
+  public int getBlockLength() {
     return length;
   }
 
-  public long getPieceIndex() {
+  public int getPieceIndex() {
     return index;
   }
 }
