+++ java/edu/ualr/bittorrent/interfaces/messages/Piece.java
@@ -3,7 +3,7 @@ package edu.ualr.bittorrent.interfaces.messages;
 import edu.ualr.bittorrent.interfaces.HasPeer;
 
 public interface Piece extends HasPeer {
-  public long getPieceIndex();
-  public long getBeginningOffset();
+  public int getPieceIndex();
+  public int getBeginningOffset();
   public byte[] getBlock();
 }
