+++ java/edu/ualr/bittorrent/interfaces/messages/Request.java
@@ -3,7 +3,7 @@ package edu.ualr.bittorrent.interfaces.messages;
 import edu.ualr.bittorrent.interfaces.HasPeer;
 
 public interface Request extends HasPeer {
-  public long getPieceIndex();
-  public long getBeginningOffset();
-  public long getBlockLength();
+  public int getPieceIndex();
+  public int getBeginningOffset();
+  public int getBlockLength();
 }
